+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-# 24 August, 2021
+# Changelog
+
+## 24 August, 2021
 
 - Initial commit
+
+## 26 August, 2021
+
+### New
+- Add project: Dynamic CSS Variables